+++ example/index.html
@@ -7,11 +7,11 @@
     <meta name="description" content="Spectrum is a JavaScript colorpicker plugin using the jQuery framework.  It is highly customizable, but can also be used as a simple input type=color polyfill">
     <meta name="author" content="Brian Grinstead and Spectrum contributors">
 
-    <link rel="stylesheet" type="text/css" href="../spectrum.css">
+    <link rel="stylesheet" type="text/css" href="../src/spectrum.css">
     <link rel="stylesheet" type="text/css" href="../docs/docs.css">
     <script type="text/javascript" src="../docs/jquery-1.9.1.js"></script>
-    <script type="text/javascript" src="../spectrum.js"></script>
-	<script type="text/javascript" src="../i18n/jquery.spectrum-fi.js"></script>
+    <script type="text/javascript" src="../src/spectrum.js"></script>
+	<script type="text/javascript" src="../src/i18n/jquery.spectrum-fi.js"></script>
     <script type='text/javascript' src='example.js'></script>
 
     <style>
