+++ index.html
@@ -7,12 +7,12 @@
     <meta name="description" content="Spectrum is a JavaScript colorpicker plugin using the jQuery framework.  It is highly customizable, but can also be used as a simple input type=color polyfill">
     <meta name="author" content="Brian Grinstead and Spectrum contributors">
 
-    <link rel="stylesheet" type="text/css" href="spectrum.css">
+    <link rel="stylesheet" type="text/css" href="dist/spectrum.css">
     <link rel="stylesheet" type="text/css" href="docs/docs.css">
     <link rel="stylesheet" type="text/css" href="docs/highlight/styles/default.css">
     <link rel="stylesheet" type="text/css" href="docs/highlight/styles/monokai-sublime.css">
     <script type="text/javascript" src="docs/jquery-1.9.1.js"></script>
-    <script type="text/javascript" src="spectrum.js"></script>
+    <script type="text/javascript" src="dist/spectrum.js"></script>
     <script type='text/javascript' src='docs/toc.js'></script>
     <script type='text/javascript' src='docs/docs.js'></script>
     <script type='text/javascript' src='docs/highlight/highlight.pack.js'></script>
