+++ src/i18n/jquery.spectrum-ar.js
