+++ src/i18n/jquery.spectrum-ca.js
