+++ src/i18n/jquery.spectrum-cs.js
