+++ src/i18n/jquery.spectrum-de.js
