+++ src/i18n/jquery.spectrum-dk.js
