+++ src/i18n/jquery.spectrum-es.js
