+++ src/i18n/jquery.spectrum-fa.js
