+++ src/i18n/jquery.spectrum-fi.js
