+++ src/i18n/jquery.spectrum-fr.js
