+++ src/i18n/jquery.spectrum-gr.js
