+++ src/i18n/jquery.spectrum-he.js
