+++ src/i18n/jquery.spectrum-hr.js
