+++ src/i18n/jquery.spectrum-id.js
