+++ src/i18n/jquery.spectrum-ja.js
