+++ src/i18n/jquery.spectrum-ko.js
