+++ src/i18n/jquery.spectrum-lt.js
