+++ src/i18n/jquery.spectrum-nl.js
