+++ src/i18n/jquery.spectrum-pl.js
