+++ src/i18n/jquery.spectrum-pt-br.js
