+++ src/i18n/jquery.spectrum-ru.js
