+++ src/i18n/jquery.spectrum-sv.js
