+++ src/i18n/jquery.spectrum-zh-cn.js
