+++ src/themes/index.html
@@ -8,9 +8,9 @@
     <meta name="author" content="Brian Grinstead and Spectrum contributors">
 
     <link rel="stylesheet" type="text/css" href="../spectrum.css">
-    <link rel="stylesheet" type="text/css" href="../docs/bootstrap.css">
-    <link rel="stylesheet" type="text/css" href="../docs/docs.css">
-    <script type="text/javascript" src="../docs/jquery-1.9.1.js"></script>
+    <link rel="stylesheet" type="text/css" href="../../docs/bootstrap.css">
+    <link rel="stylesheet" type="text/css" href="../../docs/docs.css">
+    <script type="text/javascript" src="../../docs/jquery-1.9.1.js"></script>
     <script type="text/javascript" src="../spectrum.js"></script>
