+++ test/index.html
@@ -4,12 +4,12 @@
   <meta charset="utf-8">
   <title>Spectrum Tests</title>
   <link rel="stylesheet" href="qunit.css">
-  <link rel="stylesheet" href="../spectrum.css">
+  <link rel="stylesheet" href="../src/spectrum.css">
 </head>
 <body>
   <div id="qunit"></div>
   <script src="../docs/jquery-1.9.1.js"></script>
-  <script src="../spectrum.js"></script>
+  <script src="../src/spectrum.js"></script>
   <script src="qunit.js"></script>
   <script src="tests.js"></script>
