+++ test/loaders.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <title>Spectrum Tests</title>
   <link rel="stylesheet" href="qunit.css">
-  <link rel="stylesheet" href="../spectrum.css">
+  <link rel="stylesheet" href="../src/spectrum.css">
 </head>
 <body>
   <div id="qunit"></div>
